+++ components/games/headphones/Constants.ts
@@ -1,6 +1,6 @@
 import { Dimensions } from 'react-native';
 
-const Constants = {
+export const Constants = {
     MAX_WIDTH: Dimensions.get("screen").width,
     MAX_HEIGHT: Dimensions.get("screen").height,
     CENTER: 25,
@@ -8,5 +8,3 @@ const Constants = {
     PSEUDO_PIXEL: Math.min(Dimensions.get("screen").width,Dimensions.get("screen").height)/50,
     DART_SIZE: Math.min(Dimensions.get("screen").width,Dimensions.get("screen").height)/50
 }
-
-export { Constants }
