+++ components/games/headphones/Points.js
@@ -1,0 +1,26 @@
+import React, {Component} from "react";
+import {View, Image} from "react-native";
+import { back } from "react-native/Libraries/Animated/src/Easing";
+import {Constants} from './Constants'
+
+export default class Dart extends Component{
+    constructor(props){
+        super(props);
+    }
+
+
+    render(){
+        /*let state = 
+        if (state === 2){*/
+            
+            return(
+               
+            )
+        //}
+        
+        
+        
+        
+
+    }
+}
