+++ components/games/headphones/Thread.js
@@ -1,9 +1,11 @@
 import React, {Component} from "react";
 import {View} from "react-native";
-import Svg, { Line } from "react-native-svg";
+import { back } from "react-native/Libraries/Animated/src/Easing";
+import {Constants} from './Constants';
+import Svg, { Path, Rect, Line } from "react-native-svg";
 
 export default class Node extends Component{
-    constructor(props: any){
+    constructor(props){
         super(props);
     }
